fix: address PR feedback
